backend/pyproject.toml
[project]
name = "nexus-ai-backend"
version = "0.1.0"
description = "Nexus AI — local RAG + Agent platform backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "openai>=1.50",
    "langchain>=0.3",
    "langchain-community>=0.3",
    "langchain-openai>=0.2",
    "langchain-qdrant>=0.2",
    "qdrant-client>=1.11",
    "python-multipart>=0.0.9",
    "python-jose[cryptography]>=3.3",
    "pypdf>=4.0",
    "python-docx>=1.1",
    "docx2txt>=0.8",
    "python-dotenv>=1.0",
    "pydantic>=2.8",
    "pydantic-settings>=2.5",
    "structlog>=24.0",
    "prometheus-client>=0.21",
    "prometheus-fastapi-instrumentator>=7.0",
    "httpx>=0.27",
    "duckduckgo-search>=8.0",
    "numexpr>=2.8",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
]
dev = [
    "ruff>=0.6",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
